+++ ts/features/fastLogin/backend/index.ts
@@ -1,12 +1,3 @@
-import { withoutUndefinedValues } from "@pagopa/ts-commons/lib/types";
-
-import { identity } from "fp-ts/lib/function";
-import {
-  createFetchRequestForApi,
-  ReplaceRequestParams,
-  RequestParams,
-  TypeofApiCall
-} from "@pagopa/ts-commons/lib/requests";
 import { KeyInfo } from "../../lollipop/utils/crypto";
 import { LollipopConfig } from "../../lollipop";
 import { lollipopFetch } from "../../lollipop/utils/fetch";
@@ -14,12 +5,10 @@ import { LollipopMethodEnum } from "../../../../definitions/lollipop/LollipopMet
 import { LollipopOriginalURL } from "../../../../definitions/lollipop/LollipopOriginalURL";
 import { LollipopSignatureInput } from "../../../../definitions/lollipop/LollipopSignatureInput";
 import { LollipopSignature } from "../../../../definitions/lollipop/LollipopSignature";
-import {
-  fastLoginDecoder,
-  FastLoginResponse,
-  FastLoginT
-} from "./mockedFunctionsAndTypes";
+import { defaultRetryingFetch } from "../../../utils/fetch";
+import { createClient, createLollipopClient } from "./mockedClients";
 
+// fastLogin call
 export const performFastLogin = async (fastLoginClient: FastLoginClient) =>
   await fastLoginClient.fastLogin({
     "x-pagopa-lollipop-original-method": LollipopMethodEnum.POST,
@@ -35,59 +24,20 @@ export const createFastLoginClient = (
   keyInfo: KeyInfo = {},
   lollipopConfig: LollipopConfig
 ) =>
-  createClient({
+  createLollipopClient({
     baseUrl,
-    fetchApi: lollipopFetch(lollipopConfig, keyInfo)
+    fetchApi: lollipopFetch(lollipopConfig, keyInfo, 1)
   });
 
-function createClient({
-  baseUrl,
-  fetchApi,
-  basePath = "/api/v1/fast-login"
-}: {
-  baseUrl: string;
-  fetchApi: typeof fetch;
-  basePath?: string;
-}) {
-  const options = {
+// getNonce call
+
+export const performGetNonce = async (nonceClient: GetNonceClient) =>
+  await nonceClient.getNonce({} as never);
+
+type GetNonceClient = ReturnType<typeof createNonceClient>;
+
+export const createNonceClient = (baseUrl: string) =>
+  createClient({
     baseUrl,
-    fetchApi
-  };
-
-  const fastLoginT: ReplaceRequestParams<
-    FastLoginT,
-    RequestParams<FastLoginT>
-  > = {
-    body: body => JSON.stringify(body),
-    method: "post",
-    headers: ({
-      ["x-pagopa-lollipop-original-method"]: xPagopaLollipopOriginalMethod,
-      ["x-pagopa-lollipop-original-url"]: xPagopaLollipopOriginalUrl,
-      ["signature-input"]: signatureInput,
-      ["signature"]: signature
-    }) => ({
-      "x-pagopa-lollipop-original-method": xPagopaLollipopOriginalMethod,
-
-      "x-pagopa-lollipop-original-url": xPagopaLollipopOriginalUrl,
-
-      "signature-input": signatureInput,
-
-      signature,
-
-      "Content-Type": "application/json"
-    }),
-    response_decoder: fastLoginDecoder(FastLoginResponse),
-    url: () => `${basePath}`,
-
-    query: () => withoutUndefinedValues({})
-  };
-
-  const fastLogin: TypeofApiCall<FastLoginT> = createFetchRequestForApi(
-    fastLoginT,
-    options
-  );
-
-  return {
-    fastLogin: identity(fastLogin)
-  };
-}
+    fetchApi: defaultRetryingFetch()
+  });
